server/vla_server/requirements.txt
# VLA Server Dependencies
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Fine-tuning
peft>=0.6.0
datasets>=2.14.0

# Communication
pyzmq>=25.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Optional: for faster image loading
opencv-python>=4.8.0
